+++ syspulse/gui/dialogs/__init__.py
@@ -1,0 +1,13 @@
+"""
+Dialog windows for SysPulse GUI
+
+Modal dialogs for various operations.
+"""
+
+# Dialogs will be imported here as they are implemented
+# from .clean_cache_dialog import CleanCacheDialog
+# from .cleanup_confirm import CleanupConfirmDialog
+# from .settings_dialog import SettingsDialog
+# from .about_dialog import AboutDialog
+
+__all__ = []