+++ syspulse/gui/tabs/__init__.py
@@ -1,0 +1,15 @@
+"""
+Tab widgets for SysPulse GUI
+
+Individual tab implementations for different modules.
+"""
+
+# Tab widgets will be imported here as they are implemented
+# from .dashboard import DashboardTab
+# from .browser import BrowserTab
+# from .startup import StartupTab
+# from .storage import StorageTab
+# from .processes import ProcessesTab
+# from .reports import ReportsTab
+
+__all__ = []